+++ src/main/java/com/example/ForoPrecios/controller/ProductoController.java
@@ -1,9 +1,14 @@
 package com.example.ForoPrecios.controller;
 
-import com.example.ForoPrecios.model.Producto;
+import com.example.ForoPrecios.exception.ResourceNotFoundException;
+import com.example.ForoPrecios.model.dto.ProductoDTO;
+import com.example.ForoPrecios.model.entity.Producto;
 import com.example.ForoPrecios.service.IProductoService;
+import jakarta.validation.Valid;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -19,28 +24,59 @@ public class ProductoController {
     private IProductoService productoService;
     
     @GetMapping("/productos/{id}")
-    public Producto getProducto(@PathVariable Long id){
-        return productoService.findProducto(id);
+    public ResponseEntity<?> getProducto(@PathVariable Long id){
+        Producto producto = productoService.findProducto(id);
+        if(producto == null){
+            throw new ResourceNotFoundException("Producto","id",id);
+        } 
+        return new ResponseEntity<>(producto, HttpStatus.OK);
     }
     
     @GetMapping("/productos")
-    public List<Producto> getProductos(){
-        return productoService.getProductos();
+    public ResponseEntity<?> getProductos(){
+        List<Producto> productos = productoService.getProductos(); 
+        if(productos == null || productos.isEmpty()){
+            throw new ResourceNotFoundException("Productos");
+        }
+        return new ResponseEntity<>(productos, HttpStatus.OK);
     }
     
     @PostMapping("/productos/crear")
-    public void crearProducto(@RequestBody Producto producto){
-        productoService.saveProducto(producto);
+    public ResponseEntity<?> crearProducto(@RequestBody @Valid ProductoDTO productoDTO){
+        Producto producto = Producto.builder().nombre(productoDTO.getNombre()).marca(productoDTO.getMarca()).build();
+        try{
+            productoService.saveProducto(producto);
+            return new ResponseEntity<>(producto, HttpStatus.CREATED);
+        }catch(Exception ex){
+            throw ex;
+        }
     }
     
     @PutMapping("/productos/editar/{id}")
-    public void editarProducto(@PathVariable Long id){
-        productoService.editProducto(null);
+    public ResponseEntity<?> editarProducto(@PathVariable Long id, @RequestBody @Valid ProductoDTO productoDTO){
+        Producto producto = productoService.findProducto(id);
+        if(producto == null){
+            throw new ResourceNotFoundException("Producto","id",id); 
+        }
+        producto.setId_producto(id);
+        producto.setNombre(productoDTO.getNombre());
+        producto.setMarca(productoDTO.getMarca());
+        try{
+            productoService.editProducto(producto);
+            return new ResponseEntity<>(producto, HttpStatus.OK);
+        }catch(Exception ex){
+            throw ex;
+        }
     }
     
     @DeleteMapping("/productos/eliminar/{id}")
-    public void eliminarProducto(@PathVariable Long id){
+    public ResponseEntity<?> eliminarProducto(@PathVariable Long id){
+        Producto producto = productoService.findProducto(id);
+        if(producto == null){
+            throw new ResourceNotFoundException("Producto","id",id); 
+        }
         productoService.deleteProducto(id);
+        return new ResponseEntity<>("Producto borrado exitosamente", HttpStatus.OK);
     }
     
 }
